+++ R/emuR-get_trackdata.R
@@ -49,6 +49,7 @@
 ##' a further \code{nrOfAllocationRows} more rows will be allocated. As this allocation leads to
 ##' a performance penalty one should consider increasing this number for large emuDBs. 
 ##' @param resultType Specify class of returned object. Either \code{"emuRtrackdata"} or \code{"trackdata"} (see \code{\link{trackdata}} and \code{\link{emuRtrackdata}} for details about these objects).
+##' @param persistentOutputType Prevent converting the output object to a \code{data.frame} depending on the \code{npoint} and \code{cut} arguments
 ##' @param verbose Show progress bars and further information
 ##' @return If the \code{cut} parameter is not set (the default) an object of type \code{\link{trackdata}} or \code{\link{emuRtrackdata}} 
 ##' is returned. If \code{cut} is set and \code{npoints} is not, or the seglist 
@@ -86,7 +87,7 @@
 "get_trackdata" <- function(emuDBhandle, seglist = NULL, ssffTrackName = NULL, cut = NULL, 
                             npoints = NULL, onTheFlyFunctionName = NULL, onTheFlyParams = NULL, 
                             onTheFlyOptLogFilePath = NULL, nrOfAllocationRows = 10000, 
-                            resultType = "trackdata", verbose = TRUE){
+                            resultType = "trackdata", persistentOutputType = FALSE, verbose = TRUE){
   #########################
   # get DBconfig
   DBconfig = load_DBconfig(emuDBhandle)
@@ -421,7 +422,7 @@
   data = as.matrix(data) # make sure it is a matrix to be able to set row names
   timeStampRowNames = timeStampRowNames[timeStampRowNames != -1]
   
-  if((!is.null(cut) && (npoints == 1 || is.null(npoints))) || (emusegs.type(seglist) == 'event' && (npoints == 1 || is.null(npoints)))){
+  if(!persistentOutputType && ((!is.null(cut) && (npoints == 1 || is.null(npoints))) || (emusegs.type(seglist) == 'event' && (npoints == 1 || is.null(npoints))))){
     resObj = as.data.frame(data)
     colnames(resObj) = paste(trackDef[[1]]$columnName, seq(1:ncol(resObj)), sep = '')    
   }else{